backend/requirements.txt
fastapi[standard]
uvicorn
google-generativeai
PyPDF2
pandas
duckdb
python-dotenv
python-multipart
openai